test queries moved to right class, lambda was used to filter data
